Refactored AirlockVisualizer
