+++ Content.Server/Entry/IgnoredComponents.cs
@@ -20,6 +20,7 @@ namespace Content.Server.Entry
             "PuddleVisualizer",
             "UIFragment",
             "PDABorderColor",
+            "AirlockVisuals",
         };
     }
 }
