+++ Resources/Prototypes/Entities/Structures/Doors/Airlocks/external.yml
@@ -18,8 +18,7 @@
   - type: Sprite
     sprite: Structures/Doors/Airlocks/Standard/external.rsi
   - type: Appearance
-    visuals:
-    - type: AirlockVisualizer
+  - type: AirlockVisuals
   - type: WiresVisuals
   - type: PaintableAirlock
     group: External
