+++ Resources/Prototypes/Entities/Structures/Doors/Airlocks/highsec.yml
@@ -56,8 +56,7 @@
     time: 10
   - type: Airlock
   - type: Appearance
-    visuals:
-    - type: AirlockVisualizer
+  - type: AirlockVisuals
   - type: WiresVisuals
   - type: ApcPowerReceiver
     powerLoad: 20
