+++ Resources/Prototypes/Entities/Structures/Doors/Firelocks/firelock.yml
@@ -77,10 +77,9 @@
       time: 3
     - type: Firelock
     - type: Appearance
-      visuals:
-        - type: AirlockVisualizer
-          animationTime: 0.6
-          emergencyAccessLayer: false
+    - type: AirlockVisuals
+      animationTime: 0.6
+      emergencyAccessLayer: false
     - type: WiresVisuals
     - type: Wires
       BoardName: "Firelock Control"
