+++ Resources/Prototypes/Entities/Structures/Doors/Shutter/shutters.yml
@@ -49,10 +49,9 @@
     fuel: 3
     time: 3
   - type: Appearance
-    visuals:
-      - type: AirlockVisualizer
-        simpleVisuals: true
-        animationTime: 1.4
+  - type: AirlockVisuals
+    simpleVisuals: true
+    animationTime: 1.4
   - type: UserInterface
     interfaces:
     - key: enum.WiresUiKey.Key
